Add new problems from 2020 questions Fixed entire solutions by separating projects into diff. solutions
