+++ interviewcodings2018/interviewcodings2018.csproj
@@ -4,10 +4,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F7EAE994-73CC-4A87-8AC6-8481E1BBCBB2}</ProjectGuid>
+    <ProjectGuid>{EBDB1A39-4086-4C6B-82B1-A595F8172499}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>csharpproject</RootNamespace>
-    <AssemblyName>csharpproject</AssemblyName>
+    <RootNamespace>interviewcodings2018</RootNamespace>
+    <AssemblyName>interviewcodings2018</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -43,29 +43,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AmznBlock.cs" />
-    <Compile Include="LargestConsecutiveRange.cs" />
-    <Compile Include="BalancedBrackets.cs" />
-    <Compile Include="StockPriceProfit.cs" />
-    <Compile Include="DynamicArray.cs" />
-    <Compile Include="FindDistinctNumbers.cs" />
-    <Compile Include="GivenSumSubArray.cs" />
-    <Compile Include="FindDuplicateParenthesis.cs" />
-    <Compile Include="JumpingNumbers.cs" />
-    <Compile Include="KaratsubaMultiplication.cs" />
-    <Compile Include="IProblem.cs" />
-    <Compile Include="KLargestInArray.cs" />
-    <Compile Include="LargestWordInDictionary.cs" />
-    <Compile Include="LongestPalindromeSubstring.cs" />
-    <Compile Include="LowerBoundSTL.cs" />
-    <Compile Include="MagicSquareFormation.cs" />
-    <Compile Include="MaxDiffInArray.cs" />
-    <Compile Include="PeakFinding1d.cs" />
+    <Compile Include="ArrayDuplicateChecker.cs" />
+    <Compile Include="ArrayElementSumToN.cs" />
+    <Compile Include="FindNthFibonacci.cs" />
+    <Compile Include="LongestSequenceBetweenStrings.cs" />
+    <Compile Include="GenerateRandomPointInsideCircle.cs" />
+    <Compile Include="IsStringsAnagram.cs" />
+    <Compile Include="MaximumDifferentWordInString.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SparseArray.cs" />
-    <Compile Include="TwoEqualSumSubarray.cs" />
-    <Compile Include="Utility\ArrayExtensions.cs" />
+    <Compile Include="SecondLargestElement.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
