+++ interviewcodings2018/readme.md
@@ -1,0 +1,33 @@
+﻿# The real problems asked during interview sessions
+
+### **Array Duplicate Checker**  
+Definition: 
+Solution: [ArrayDuplicateChecker.cs](ArrayDuplicateChecker.cs)
+
+### **Array Element sum to N**  
+Definition: 
+Solution: [ArrayElementSumToN.cs](ArrayElementSumToN.cs)
+
+### **Find nth fibonacci**  
+Definition: 
+Solution: [FindNthFibonacci.cs](FindNthFibonacci.cs)
+
+### **Generate random point inside circle**  
+Definition: 
+Solution: [GenerateRandomPointInsideCircle.cs](GenerateRandomPointInsideCircle.cs)
+
+### **Check if given string is Anagram**  
+Definition: 
+Solution: [IsStringsAnagram.cs](IsStringsAnagram.cs)
+
+### **Find longest sequence between strings**  
+Definition: 
+Solution: [LongestSequenceBetweenStrings.cs](LongestSequenceBetweenStrings.cs)
+
+### **Find maximum difference word in given string**  
+Definition: 
+Solution: [MaximumDifferentWordInString.cs](MaximumDifferentWordInString.cs)
+
+### **Find second largest element in given array**  
+Definition: 
+Solution: [SecondLargestElement.cs](SecondLargestElement.cs)
