+++ interviewcodings2020/IProblem.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace interviewcodings2020
+{
+    interface IProblem
+    {
+        void Execute();
+
+    }
+}
