+++ interviewcodings2020/interviewcodings2020.csproj
@@ -4,10 +4,10 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F7EAE994-73CC-4A87-8AC6-8481E1BBCBB2}</ProjectGuid>
+    <ProjectGuid>{BA6D3311-7753-42FA-BAB5-A2F71AB84397}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>csharpproject</RootNamespace>
-    <AssemblyName>csharpproject</AssemblyName>
+    <RootNamespace>interviewcodings2020</RootNamespace>
+    <AssemblyName>interviewcodings2020</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -43,29 +43,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AmznBlock.cs" />
-    <Compile Include="LargestConsecutiveRange.cs" />
-    <Compile Include="BalancedBrackets.cs" />
-    <Compile Include="StockPriceProfit.cs" />
-    <Compile Include="DynamicArray.cs" />
-    <Compile Include="FindDistinctNumbers.cs" />
-    <Compile Include="GivenSumSubArray.cs" />
-    <Compile Include="FindDuplicateParenthesis.cs" />
-    <Compile Include="JumpingNumbers.cs" />
-    <Compile Include="KaratsubaMultiplication.cs" />
+    <Compile Include="GoatLatin.cs" />
+    <Compile Include="InsertDigitToReturnMax.cs" />
     <Compile Include="IProblem.cs" />
-    <Compile Include="KLargestInArray.cs" />
-    <Compile Include="LargestWordInDictionary.cs" />
-    <Compile Include="LongestPalindromeSubstring.cs" />
-    <Compile Include="LowerBoundSTL.cs" />
-    <Compile Include="MagicSquareFormation.cs" />
-    <Compile Include="MaxDiffInArray.cs" />
-    <Compile Include="PeakFinding1d.cs" />
+    <Compile Include="PalindromeAlphaCharacters.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SparseArray.cs" />
-    <Compile Include="TwoEqualSumSubarray.cs" />
-    <Compile Include="Utility\ArrayExtensions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
