--- interviewcodingsunittests/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MSTest.TestAdapter" version="2.1.1" targetFramework="net48" />
-  <package id="MSTest.TestFramework" version="2.1.1" targetFramework="net48" />
-  <package id="NUnit" version="3.12.0" targetFramework="net48" />
-</packages>
