+++ otherkata.sln
@@ -1,0 +1,65 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30320.27
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "oopfundamentals", "oopfundamentals\oopfundamentals.csproj", "{42B718AC-9306-4106-AF31-A7AA7C87EB75}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JonahRuleEngine", "JonahRuleEngine\JonahRuleEngine.csproj", "{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dsaexposed", "dsaexposed\dsaexposed.csproj", "{949C3854-17E0-459A-AEA4-0FBAB037299B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|x64.Build.0 = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Debug|x86.Build.0 = Debug|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|x64.ActiveCfg = Release|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|x64.Build.0 = Release|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|x86.ActiveCfg = Release|Any CPU
+		{42B718AC-9306-4106-AF31-A7AA7C87EB75}.Release|x86.Build.0 = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|x64.Build.0 = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Debug|x86.Build.0 = Debug|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|x64.ActiveCfg = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|x64.Build.0 = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|x86.ActiveCfg = Release|Any CPU
+		{1CC9D639-23BB-4A9A-86B8-BAE7573CAFBD}.Release|x86.Build.0 = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|x64.Build.0 = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Debug|x86.Build.0 = Debug|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|x64.ActiveCfg = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|x64.Build.0 = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|x86.ActiveCfg = Release|Any CPU
+		{949C3854-17E0-459A-AEA4-0FBAB037299B}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {632162A0-3FC6-401C-BD55-AA97842DB74E}
+	EndGlobalSection
+EndGlobal
