--- rsedgewickalgos/readme.md
@@ -1,14 +1,0 @@
- 
-# Algorithms: 24-part Lecture Series
-This section includes all the problems discussed from the book: Robert Sedgewick / Kevin Wayne  
-Book Url: [Algorithms - 4th Edition](https://algs4.cs.princeton.edu/home/)  
-
-
-### **ArrayTripletsSumToZero**
-Definition:
-```
-Given N distinct integers, how many triples sum to exactly zero?    
-```
-Reference: [Chapter2](https://algs4.cs.princeton.edu/14analysis/)
-Solution: WIP  
-
