+++ testprojects/codekatatest.sln
@@ -1,0 +1,51 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "interviewcodings2018unittests", "interviewcodings2018unittests\interviewcodings2018unittests.csproj", "{8BAD4E12-A576-4207-AA8C-30810A2C69E1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "interviewcodings2018", "..\interviewcodings2018\interviewcodings2018.csproj", "{EBDB1A39-4086-4C6B-82B1-A595F8172499}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|x64.Build.0 = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Debug|x86.Build.0 = Debug|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|x64.ActiveCfg = Release|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|x64.Build.0 = Release|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|x86.ActiveCfg = Release|Any CPU
+		{8BAD4E12-A576-4207-AA8C-30810A2C69E1}.Release|x86.Build.0 = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|x64.Build.0 = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Debug|x86.Build.0 = Debug|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|x64.ActiveCfg = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|x64.Build.0 = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|x86.ActiveCfg = Release|Any CPU
+		{EBDB1A39-4086-4C6B-82B1-A595F8172499}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BFB440C9-4EEA-4B6A-9D9E-6EC6BBB31DCA}
+	EndGlobalSection
+EndGlobal
