+++ testprojects/interviewcodings2018unittests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestAdapter" version="1.3.2" targetFramework="net48" />
+  <package id="MSTest.TestFramework" version="1.3.2" targetFramework="net48" />
+</packages>
